+++ config/routes.php
@@ -186,7 +186,7 @@ Router::scope('/', ['plugin' => 'MeCms'], function ($routes) {
 		/**
 		 * Other admin routes
 		 */
-		$controllers = ['backups', 'banners', 'banners_positions', 'pages', 'photos_albums', 'photos', 'posts_categories', 'posts_tags', 'posts', 'systems', 'tags', 'users', 'users_groups'];
+		$controllers = ['backups', 'banners', 'banners_positions', 'logs', 'pages', 'photos_albums', 'photos', 'posts_categories', 'posts_tags', 'posts', 'systems', 'tags', 'users', 'users_groups'];
 		$controllers = sprintf('(%s)', implode('|', $controllers));
 		
 		$routes->connect('/:controller', [], ['controller' => $controllers]);